pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["farzadgolzari"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
annotated-types = "0.7.0"
anyio = "4.6.2.post1"
astroid = "3.3.5"
certifi = "2024.8.30"
charset-normalizer = "3.4.0"
click = "8.1.7"
colorama = "0.4.6"
contourpy = "1.3.1"
cycler = "0.12.1"
dill = "0.3.9"
fastapi = "0.115.5"
filelock = "3.16.1"
fonttools = "4.55.0"
fsspec = "2024.10.0"
greenlet = "3.1.1"
h11 = "0.14.0"
idna = "3.10"
isort = "5.13.2"
jinja2 = "3.1.4"
kiwisolver = "1.4.7"
markupsafe = "3.0.2"
matplotlib = "3.9.2"
mccabe = "0.7.0"
mpmath = "1.3.0"
networkx = "3.4.2"
numpy = "1.26.4"
nvidia-cublas-cu12 = "12.4.5.8"
nvidia-cuda-cupti-cu12 = "12.4.127"
nvidia-cuda-nvrtc-cu12 = "12.4.127"
nvidia-cuda-runtime-cu12 = "12.4.127"
nvidia-cudnn-cu12 = "9.1.0.70"
nvidia-cufft-cu12 = "11.2.1.3"
nvidia-curand-cu12 = "10.3.5.147"
nvidia-cusolver-cu12 = "11.6.1.9"
nvidia-cusparse-cu12 = "12.3.1.170"
nvidia-nccl-cu12 = "2.21.5"
nvidia-nvjitlink-cu12 = "12.4.127"
nvidia-nvtx-cu12 = "12.4.127"
opencv-python = "4.10.0.84"
packaging = "24.2"
pandas = "2.2.3"
pillow = "11.0.0"
platformdirs = "4.3.6"
psutil = "6.1.0"
py-cpuinfo = "9.0.0"
pydantic = "2.9.2"
pydantic-core = "2.23.4"
pylint = "3.3.1"
pyparsing = "3.2.0"
python-dateutil = "2.9.0.post0"
python-multipart = "0.0.17"
pytz = "2024.2"
pyyaml = "6.0.2"
requests = "2.32.3"
scipy = "1.14.1"
seaborn = "0.13.2"
setuptools = "75.5.0"
six = "1.16.0"
sniffio = "1.3.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
starlette = "0.41.2"
sympy = "1.13.1"
tomlkit = "0.13.2"
torch = "2.5.1"
torchvision = "0.20.1"
tqdm = "4.67.0"
triton = "3.1.0"
typing-extensions = "4.12.2"
tzdata = "2024.2"
ultralytics = "8.3.33"
ultralytics-thop = "2.0.11"
urllib3 = "2.2.3"
uvicorn = "0.32.0"
asyncmy = "^0.2.9"
cryptography = "^44.0.0"
aiomysql = "^0.2.0"
xlrd = "^2.0.1"
openpyxl = "^3.1.5"
passlib = "^1.7.4"
python-dotenv = "^1.0.1"
pyjwt = "^2.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
